Split runtime snapshot coverage modules
